feat: created prod build for hydrofabric data object versioning

+++ src/icefabric/schemas/iceberg_tables/hydrofabric_snapshots.py
@@ -2,7 +2,7 @@
 
 import pyarrow as pa
 from pyiceberg.schema import Schema
-from pyiceberg.types import IntegerType, NestedField
+from pyiceberg.types import LongType, NestedField
 
 
 class HydrofabricSnapshot:
@@ -54,16 +54,16 @@ class HydrofabricSnapshot:
             PyIceberg schema for Hydrofabric
         """
         return Schema(
-            NestedField(1, "divide-attributes", IntegerType(), required=False),
-            NestedField(2, "divides", IntegerType(), required=False),
-            NestedField(3, "flowpath-attributes", IntegerType(), required=False),
-            NestedField(4, "flowpath-attributes-ml", IntegerType(), required=False),
-            NestedField(5, "flowpaths", IntegerType(), required=False),
-            NestedField(6, "hydrolocations", IntegerType(), required=False),
-            NestedField(7, "lakes", IntegerType(), required=False),
-            NestedField(8, "network", IntegerType(), required=False),
-            NestedField(9, "nexus", IntegerType(), required=False),
-            NestedField(10, "pois", IntegerType(), required=False),
+            NestedField(1, "divide-attributes", LongType(), required=False),
+            NestedField(2, "divides", LongType(), required=False),
+            NestedField(3, "flowpath-attributes", LongType(), required=False),
+            NestedField(4, "flowpath-attributes-ml", LongType(), required=False),
+            NestedField(5, "flowpaths", LongType(), required=False),
+            NestedField(6, "hydrolocations", LongType(), required=False),
+            NestedField(7, "lakes", LongType(), required=False),
+            NestedField(8, "network", LongType(), required=False),
+            NestedField(9, "nexus", LongType(), required=False),
+            NestedField(10, "pois", LongType(), required=False),
         )
 
     @classmethod
@@ -77,15 +77,15 @@ class HydrofabricSnapshot:
         """
         return pa.schema(
             [
-                pa.field("divide-attributes", pa.int32(), nullable=True),
-                pa.field("divides", pa.int32(), nullable=True),
-                pa.field("flowpath-attributes", pa.int32(), nullable=True),
-                pa.field("flowpath-attributes-ml", pa.int32(), nullable=True),
-                pa.field("flowpaths", pa.int32(), nullable=True),
-                pa.field("hydrolocations", pa.int32(), nullable=True),
-                pa.field("lakes", pa.int32(), nullable=True),
-                pa.field("network", pa.int32(), nullable=True),
-                pa.field("nexus", pa.int32(), nullable=True),
-                pa.field("pois", pa.int32(), nullable=True),
+                pa.field("divide-attributes", pa.int64(), nullable=True),
+                pa.field("divides", pa.int64(), nullable=True),
+                pa.field("flowpath-attributes", pa.int64(), nullable=True),
+                pa.field("flowpath-attributes-ml", pa.int64(), nullable=True),
+                pa.field("flowpaths", pa.int64(), nullable=True),
+                pa.field("hydrolocations", pa.int64(), nullable=True),
+                pa.field("lakes", pa.int64(), nullable=True),
+                pa.field("network", pa.int64(), nullable=True),
+                pa.field("nexus", pa.int64(), nullable=True),
+                pa.field("pois", pa.int64(), nullable=True),
             ]
         )
